Test: trng: Remove dependency on mbedTLS
This test builds/runs without it. Nor mbedTLS is among dependencies of
MCU-Driver-HAL. As MCU-Driver-HAL is intended to be a hardware interface
library, it would be strange to have it depend on mbedTLS. Expectedly,
it would be the opposite - mbedTLS could include support for
MCU-Driver-HAL to provide cross-device support for TRNG service.

Signed-off-by: Paul Sokolovsky <paul.sokolovsky@linaro.org>